match: CGameObject::Notify is a compound -=; Scan's dwell gate is unsigned
Notify 84.48 -> 100 EXACT. Retail loads the subtrahend first
(`mov edx,[eax+0x120]` before `mov eax,[ecx+0x128]`) and stores between
the test and the branch - that evaluation order is `m_health -= p->m_damage`,
not `i32 d = m_health - p->m_damage; m_health = d;`.

CBattlezMapConfig::Scan 71.80 -> 72.30: retail's dwell gate is `jbe`, so
the m_dwell/m_inactiveTargetRerouteDelay comparison is unsigned.

+++ src/Gruntz/TileScan.cpp
@@ -19,7 +19,7 @@ static inline i32 GridLookup(CMapMgr* g, i32 x, i32 y) {
 // @early-stop
 RVA(0x00035f10, 0x155)
 i32 CBattlezMapConfig::Scan(CGrunt* arg) {
-    if (arg->m_dwell <= m_inactiveTargetRerouteDelay) {
+    if (static_cast<u32>(arg->m_dwell) <= static_cast<u32>(m_inactiveTargetRerouteDelay)) {
         return 1;
     }
     i32 v = arg->m_targetTeam;
